backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
httpx==0.24.1
python-dotenv==1.0.0
aiofiles==23.2.1
python-multipart==0.0.6
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.1.0
supabase==2.0.0
